Removed unnecessary logic for calculated columns
