+++ modules/engines/engine-hdb/src/main/java/com/sap/xsk/hdb/ds/transformer/hdbdd/HdbddTransformer.java
@@ -12,9 +12,7 @@
 package com.sap.xsk.hdb.ds.transformer.hdbdd;
 
 import com.sap.xsk.hdb.ds.model.XSKDBContentType;
-import com.sap.xsk.hdb.ds.model.hdbtable.XSKDataStructureHDBTableCalculatedColumnModel;
 import com.sap.xsk.hdb.ds.model.hdbtable.XSKDataStructureHDBTableColumnModel;
-import com.sap.xsk.hdb.ds.model.hdbtable.XSKDataStructureHDBTableCommonColumnModel;
 import com.sap.xsk.hdb.ds.model.hdbtable.XSKDataStructureHDBTableConstraintForeignKeyModel;
 import com.sap.xsk.hdb.ds.model.hdbtable.XSKDataStructureHDBTableConstraintPrimaryKeyModel;
 import com.sap.xsk.hdb.ds.model.hdbtable.XSKDataStructureHDBTableConstraintUniqueModel;
@@ -80,7 +78,6 @@ public class HdbddTransformer {
     primaryKey.setName("PK_" + tableModel.getName());
     tableModel.getConstraints().setPrimaryKey(primaryKey);
 
-    List<XSKDataStructureHDBTableCalculatedColumnModel> calculatedColumns = new ArrayList<>();
     List<XSKDataStructureHDBTableColumnModel> tableColumns = new ArrayList<>();
     for(EntityElementSymbol currentElement: entitySymbol.getElements()) {
       if (currentElement.getType() instanceof StructuredDataTypeSymbol) {
@@ -88,8 +85,6 @@ public class HdbddTransformer {
         subElements.forEach(subE -> {
           tableColumns.add(transformFieldSymbolToColumnModel(subE, true));
         });
-      } else if(currentElement.isCalculatedColumn()) {
-        calculatedColumns.add(transformFieldSymbolToCalculatedColumnModel(currentElement));
       } else {
         tableColumns.add(transformFieldSymbolToColumnModel(currentElement, true));
       }
@@ -133,7 +128,6 @@ public class HdbddTransformer {
     });
 
     tableModel.setColumns(tableColumns);
-    tableModel.setCalculatedColumns(calculatedColumns);
     tableModel.setLocation(location);
     if (entitySymbol.getAnnotation(CATALOG_ANNOTATION) != null) {
       String tableType = entitySymbol.getAnnotation(CATALOG_ANNOTATION).getKeyValuePairs().get(CATALOG_OBJ_TABLE_TYPE).getValue();
@@ -356,6 +350,8 @@ public class HdbddTransformer {
         columnModel.setPrimaryKey(elementSymbol.isKey());
       }
 
+      columnModel.setCalculatedColumn(elementSymbol.isCalculatedColumn());
+      columnModel.setStatement(elementSymbol.getStatement());
       columnModel.setNullable(!elementSymbol.isNotNull());
       columnModel.setDefaultValue(elementSymbol.getDefaultValue());
       columnModel.setDefaultValueDateTimeFunction(elementSymbol.isDefaultValueDateTimeFunction());
@@ -383,41 +379,6 @@ public class HdbddTransformer {
     return columnModel;
   }
 
-  /**
-   * @param fieldSymbol: fieldSymbol
-   */
-  private XSKDataStructureHDBTableCalculatedColumnModel transformFieldSymbolToCalculatedColumnModel(FieldSymbol fieldSymbol) {
-    XSKDataStructureHDBTableCalculatedColumnModel calculatedColumnModel = new XSKDataStructureHDBTableCalculatedColumnModel();
-
-    calculatedColumnModel.setColumnName(fieldSymbol.getName());
-
-    if (fieldSymbol instanceof EntityElementSymbol) {
-      EntityElementSymbol elementSymbol = (EntityElementSymbol) fieldSymbol;
-      calculatedColumnModel.setStatement(elementSymbol.getStatement());
-    }
-
-    if (fieldSymbol.getType() instanceof BuiltInTypeSymbol) {
-      BuiltInTypeSymbol builtInTypeSymbol = (BuiltInTypeSymbol) fieldSymbol.getType();
-      if (builtInTypeSymbol.isHanaType()) {
-        setHanaType(calculatedColumnModel, builtInTypeSymbol);
-      } else {
-        setSqlType(calculatedColumnModel, builtInTypeSymbol);
-      }
-    } else if (fieldSymbol.getType() instanceof DataTypeSymbol) {
-      DataTypeSymbol dataType = (DataTypeSymbol) fieldSymbol.getType();
-      if (!(dataType.getType() instanceof StructuredDataTypeSymbol)) {
-        BuiltInTypeSymbol builtInType = (BuiltInTypeSymbol) dataType.getType();
-        setSqlType(calculatedColumnModel, builtInType);
-      } else {
-        StructuredDataTypeSymbol structuredDataTypeSymbol = (StructuredDataTypeSymbol) dataType.getType();
-        transformStructuredDataTypeToHdbTableType(structuredDataTypeSymbol);
-      }
-    }
-
-    return calculatedColumnModel;
-  }
-
-
   private List<XSKDataStructureHDBTableColumnModel> transformAssociationToColumnModels(AssociationSymbol associationSymbol) {
     List<XSKDataStructureHDBTableColumnModel> tableColumns = new ArrayList<>();
     associationSymbol.getForeignKeys().forEach(fk -> {
@@ -432,7 +393,7 @@ public class HdbddTransformer {
     return tableColumns;
   }
 
-  private void setSqlType(XSKDataStructureHDBTableCommonColumnModel columnModel, BuiltInTypeSymbol builtInTypeSymbol) {
+  private void setSqlType(XSKDataStructureHDBTableColumnModel columnModel, BuiltInTypeSymbol builtInTypeSymbol) {
     String typeName = builtInTypeSymbol.getName();
     CdsTypeEnum cdsTypeEnum = CdsTypeEnum.valueOf(typeName);
 
@@ -446,7 +407,7 @@ public class HdbddTransformer {
     columnModel.setType(cdsTypeEnum.getSqlType());
   }
 
-  private void setHanaType(XSKDataStructureHDBTableCommonColumnModel columnModel, BuiltInTypeSymbol builtInTypeSymbol) {
+  private void setHanaType(XSKDataStructureHDBTableColumnModel columnModel, BuiltInTypeSymbol builtInTypeSymbol) {
     String typeName = builtInTypeSymbol.getName();
     CdsHanaTypeEnum cdsHanaTypeEnum = CdsHanaTypeEnum.valueOf(typeName);
